Fixed KalViewController delegate/dataSource retain bug.
A regression was introduced in SHA: 8fa33f42 where the
KalViewController's delegate and dataSource properties were being
retained by the setter implementation but were declared to be assigned
properties. My intent was that these were weak references, and so this
commit removes the retains. As a client of KalViewController, it is
your responsibility to ensure that the lifetime of the
KalViewController's delegate and dataSource properties has at least the
same extent as the KalViewController object itself.

This closes issue #18 on GitHub.