Made it easier to override KalViewController's title.
Prior to this commit, KalViewController used to clobber its 'title'
property inherited from UIViewController in its loadView method. I
changed this so that now it only sets a default title if the title
property is nil. I modified both of the example apps to set the
KalViewController's title explicitly.

This makes it easier for clients of KalViewController to customize its
behavior without resorting to subclassing or the even more drastic
measure of modifying KalViewController's implementation.

This closes issue #17 on GitHub.